# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Курсовая работа по предмету \"Информационные системы и базы данных\" студента группы P34312, Соловьева Павла",
        "title": "Автомобилестроение в экономике",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
        "/v1/accept_subsidy": {
            "post": {
                "description": "Accept and link subsidy",
                "tags": [
                    "Posts"
                ],
                "summary": "accept subsidy",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pahan_internal_controller_http_v1.createAcceptSubsidyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pahan_internal_controller_http_v1.errResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/create_subsidy": {
            "post": {
                "description": "Create and link subsidy with dependent values",
                "tags": [
                    "Posts"
                ],
                "summary": "create subsidy",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pahan_internal_controller_http_v1.createSubsidyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pahan_internal_controller_http_v1.errResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/get_subsidies": {
            "get": {
                "description": "Get all subsidies",
                "tags": [
                    "Gets"
                ],
                "summary": "list of subsidies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Subsidy"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pahan_internal_controller_http_v1.errResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Subsidy": {
            "type": "object",
            "properties": {
                "country_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "require_price": {
                    "type": "integer"
                },
                "required_wd": {
                    "type": "string"
                }
            }
        },
...
        "pahan_internal_controller_http_v1.createAcceptSubsidyRequest": {
            "type": "object",
            "properties": {
                "component-bumper-id": {
                    "type": "integer"
                },
                "component-door-id": {
                    "type": "integer"
                },
                "component-engine-id": {
                    "type": "integer"
                },
                "component-transmission-id": {
                    "type": "integer"
                },
                "engineer-id": {
                    "type": "integer"
                },
                "factory-id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "significance": {
                    "type": "integer"
                },
                "subsidy-id": {
                    "type": "integer"
                },
                "vendor-id": {
                    "type": "integer"
                }
            }
        },
...
        "pahan_internal_controller_http_v1.createSubsidyRequest": {
            "type": "object",
            "properties": {
                "country-id-by": {
                    "type": "integer"
                },
                "require-price-by": {
                    "type": "number"
                },
                "required-wd-by": {
                    "type": "string"
                }
            }
        },
        "pahan_internal_controller_http_v1.errResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}